services/stats-dashboard/requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5
Werkzeug==3.1.5
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
Flask-Migrate==4.0.5
pdfplumber==0.11.9
pdfminer.six==20251230
Pillow==12.1.0
